travis: Drop extra coverage report regeration.
Assume we have a more recent version (3 or later) of pytest
which can be invoked as such. Also remove the report generation
switches. The codecov.io docs don't suggest this, so I don't
think it's necessary for upload to the site. There's no point
generating the reports in ci if they're not uploaded anywhere.