Mark parser patterns as raw strings.
pycodestyle/pep8 warns about invalid escape sequences in most
of the ply.lex re patterns. Avoid this by marking them as raw
strings, which means the python interpreter won't process the
backslash-escapes itself but pass them as-is to the regex engine.

This was already the case in a number of places; this just
makes the use of raw strings more uniquitous.

I'm not sure how to do the interaction with \uXXXX escapes, which
warn either way. I've left them as they were so there are still
a handful of warnings.