Always run slow tests.
Remove the --runslow option from the pytest config and instead
always run the slow tests. Unless a large corpus is passed
through the `oracc_corpus_path` environment variable, this
only adds the included sample corpus. That adds about 60%
to the pytest runtime, but is still under 10 seconds.

The motivation for this is that the way the option was accessed
is deprecated in pytest 4.1.

  https://docs.pytest.org/en/latest/deprecations.html#pytest-config-global

Having less configuration is less code to worry about, and less
variance between different test invocations, so removing the
option to skip the sample corpus is the most straightforward fix.