Don't skip 0 values when serializing to json.
By default we skip serializing properties which test as False,
to avoid cluttering the output with a lot of empty arrays.
Booleans were excluded from this, since a property which is
literally False still carries useful information.

Also exclude numbers from this check, so such properties
don't disappear if they happen to be zero-valued, which is
equally surprising.

Since `bool` is a subclass of `int` checking against
`numbers.Number` includes both cases.