test_atfserializer: Convert to pytest.
This set of tests were written to use the `unittest` framework
from the standard library but we were invoking them through pytest.

Remove the class and promote its methods to module-level functions.
The setUp method wasn't actually used, and self was only used to
lookup staticmethods, so there are otherwise few changes.