test_atfserializer: Remove api wrapper methods.
Since we call the parser and serializer with the default parameters,
there's not much value to wrapping the actual API calls this way.

The prefix/imperative style may be slightly easier to read than
the suffix/functional style, but I prefer making explicit what
functions are being called in the test.