test_atfparser: Covert to pytest.
This set of tests were written to use the `unittest` framework
from the standard library but we invoke them through pytest,
which provides better reporting.

Remove the unittest dependency here to align with changes made
to other tests.

As a minimal change, keep the TestParser class, but remove the
inheritance from `unittest.TestCase`. The pytest package instead
finds tests based on general introspection. By marking the setUp
method, which has a special meaning in the unittest framework as
a pytest fixture with `autouse=True` we ensure it is called before
every test to set up the member variable.