test_atfparser: Promote tests to module level.
Move the lexer fixture into the `try_parse` helper. None of the
tests need subsequent access to the lexer state, so there's no
reason for it to be shared.

Without shared state, the TestParser class can itself be removed.
It was necessary under the `unittest` framework, but `pytest` can
use module functions instead. This simplifies the code and saves
an indent level.